BE sửa đếm like

+++ src/main/java/com/example/social_network/service/ILikeService.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface ILikeService {
     List<Users> listLikePost(Long idPost);
     List<Users> listLikeComment(Long idComment);
-    int countLikePost(Long idPost);
-    int countLikeComment(Long Comment);
+    Long countLikePost(Long idPost);
+    Long countLikeComment(Long Comment);
     void createLikePost(Long idUser,Long idPost);
     void createLikeComment(Long idUser,Long idComment);
     void deleteLike(Long idLike);
